Making tests follow test-table convention
